🐛 : Fix bugs (missing translation and decimal for story points fields)
- http://agilemind.tech/story/edit/3
- http://agilemind.tech/story/edit/7

+++ src/main/java/com/knife/agilemind/domain/story/StoryEntity.java
@@ -52,7 +52,7 @@ public class StoryEntity implements Serializable {
      * The story points
      */
     @Column(name = "points")
-    private Long points;
+    private Double points;
 
     /**
      * The business value
@@ -164,7 +164,7 @@ public class StoryEntity implements Serializable {
      *
      * @return points
      */
-    public Long getPoints() {
+    public Double getPoints() {
         return points;
     }
 
@@ -175,7 +175,7 @@ public class StoryEntity implements Serializable {
      *
      * @return this
      */
-    public StoryEntity setPoints(Long points) {
+    public StoryEntity setPoints(Double points) {
         this.points = points;
         return this;
     }
